lidar/requirements.txt
pyserial==3.5
rplidar-roboticia==0.9.5
